Update ElevatorGUI.py
Add preset elevator levels that can be reassigned on the fly.
Travel by steps instead of inches to avoid inaccurate measurements from the conversion between the two.
Each microstepping mode reduces the number of steps taken by a factor of the reciprocal of the mode. Account for this and step/track the appropriate amount of steps.
Add option to adjust torque by changing the ratio of when the pulse is high and when it is low (while keeping the period of pulse constant; this period can be controlled using the speed input).
Add a window that displays command information for convenience.